web,desktop: factor out progress indication into common code
