+++ desktop/src/pages/home.tsx
@@ -3,14 +3,13 @@ import {
   PrimaryButton,
   SecondaryButton,
 } from 'transcribee-ui-common/components/button';
-import { Tooltip } from 'transcribee-ui-common/components/tooltip';
+import {
+  calculateTranscriptionProgress,
+  TranscriptionProgressIndicator,
+} from 'transcribee-ui-common/components/transcription_progress';
 import { invoke } from '@tauri-apps/api/core';
 import { useTauriState } from '../util/use_tauri_event';
 import { IoClose } from 'react-icons/io5';
-import { FaRegClock } from 'react-icons/fa6';
-import { IoIosCloseCircle } from 'react-icons/io';
-import { clsx } from 'clsx';
-import { ComponentProps } from 'react';
 import { Menu, MenuItem } from '@tauri-apps/api/menu';
 import { ask, message } from '@tauri-apps/plugin-dialog';
 import { getAllWindows } from '@tauri-apps/api/window';
@@ -107,67 +106,15 @@ function TranscriptionQueue({ documents }: { documents: Document[] }) {
       <h2 className="font-semibold text-center px-1 mb-4">Transcription Queue</h2>
       <div>
         {documents.map((doc) => {
-          let progressIndicator;
           let contextMenuDeleteText;
           if (doc.tasks.some((t) => t.state == 'ABORTED' || t.state == 'FAILED')) {
             contextMenuDeleteText = 'Remove Incomplete Document';
-            progressIndicator = (
-              <Tooltip
-                placement={'right'}
-                fallbackPlacements={['bottom', 'top']}
-                tooltipText={
-                  <span>
-                    Automatic transcription did not complete! This may happen <br />
-                    when transcribee is closed while a task is running.
-                  </span>
-                }
-                className="ml-2"
-              >
-                <IoIosCloseCircle className="text-red-600 shrink-0" size={21} />
-              </Tooltip>
-            );
           } else if (doc.tasks.every((t) => t.state == 'NEW')) {
-            contextMenuDeleteText = 'Remove from Transcription Queue';
-            progressIndicator = (
-              <Tooltip
-                placement={'right'}
-                fallbackPlacements={['bottom', 'top']}
-                tooltipText={
-                  <span>
-                    in queue <br />
-                    (not started yet)
-                  </span>
-                }
-                className="ml-2"
-              >
-                <FaRegClock className="text-neutral-400 shrink-0" size={21} />
-              </Tooltip>
-            );
+            contextMenuDeleteText = 'Remove From Transcription Queue';
+          } else if (doc.tasks.every((t) => t.state == 'COMPLETED')) {
+            contextMenuDeleteText = 'Delete Transcribed Document';
           } else {
-            const transcriptionProgress = calculateTranscriptionProgress(doc.tasks);
-            if (transcriptionProgress == 1) {
-              contextMenuDeleteText = 'Delete Transcribed Document';
-            } else {
-              contextMenuDeleteText = 'Abort Transcription and Delete';
-            }
-            progressIndicator = (
-              <Tooltip
-                placement={'right'}
-                fallbackPlacements={['bottom', 'top']}
-                tooltipText={
-                  transcriptionProgress == 1
-                    ? `transcription done`
-                    : `transcription ${(transcriptionProgress * 100).toFixed(0)}%`
-                }
-                className="ml-2"
-              >
-                <ProgressPie
-                  progress={transcriptionProgress}
-                  lineWidth={0.25}
-                  className="w-[21px] shrink-0"
-                />
-              </Tooltip>
-            );
+            contextMenuDeleteText = 'Abort Transcription and Delete';
           }
           return (
             <div
@@ -213,95 +160,13 @@ function TranscriptionQueue({ documents }: { documents: Document[] }) {
               <div title={doc.save_path} className="truncate text-sm">
                 {doc.display_name}
               </div>
-              {progressIndicator}
+              <TranscriptionProgressIndicator tasks={doc.tasks} />
             </div>
           );
         })}
       </div>
     </div>
   );
-}
-
-function ProgressPie({
-  progress,
-  lineWidth,
-  className,
-  ...props
-}: { progress: number; lineWidth: number } & ComponentProps<'svg'>) {
-  const progressReal = Math.min(Math.max(progress, 0.1), 0.9999);
-  const endX = Math.cos(progressReal * 2 * Math.PI - Math.PI / 2);
-  const endY = Math.sin(progressReal * 2 * Math.PI - Math.PI / 2);
-
-  const minXY = -1 - lineWidth;
-  const wh = 2 + lineWidth * 2;
-
-  return (
-    <svg
-      className={clsx(
-        className,
-        progress >= 1 ? 'text-green-600 dark:text-green-300' : 'text-black dark:text-white',
-      )}
-      viewBox={`${minXY} ${minXY} ${wh} ${wh}`}
-      stroke="currentColor"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      strokeWidth={lineWidth}
-      fill="none"
-      {...props}
-    >
-      <path
-        className="animate-[spin_3s_linear_infinite]"
-        d={
-          // circle segment according to progress
-          `
-          M ${endX} ${endY}
-          A 1 1 0 ${progressReal > 0.5 ? 1 : 0} 0 0 -1
-        `
-        }
-      />
-
-      <circle r="1" opacity={0.1} />
-
-      {progress >= 1 && (
-        // checkbox:
-        <path
-          d={`
-          M -0.42 0.1
-          L -0.12 0.4
-          L 0.47 -0.3
-        `}
-        />
-      )}
-    </svg>
-  );
-}
-
-export function calculateTranscriptionProgress(
-  tasks: {
-    task_type: 'IDENTIFY_SPEAKERS' | 'TRANSCRIBE' | 'REENCODE';
-    state: 'NEW' | 'ASSIGNED' | 'COMPLETED' | 'FAILED' | 'ABORTED';
-    current_attempt: { progress: number | null } | null;
-  }[],
-) {
-  const weights = {
-    IDENTIFY_SPEAKERS: 0.1,
-    TRANSCRIBE: 2.0,
-    REENCODE: 0.1,
-  };
-  let numerator = 0;
-  let denominator = 0;
-  for (const task of tasks) {
-    const weight = weights[task.task_type];
-    let progress = 0.0;
-    if (task.current_attempt && task.current_attempt.progress) {
-      progress = task.current_attempt.progress;
-    } else if (task.state == 'COMPLETED') {
-      progress = 1.0;
-    }
-    numerator += weight * progress;
-    denominator += weight;
-  }
-  return numerator / denominator;
 }
 
 function RecentDocuments({ documents }: { documents: Document[] }) {
